set up react-router
