+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router": "^7.1.5",
     "uuid": "^11.0.5"
   },
   "devDependencies": {
