--- src/App.css
@@ -1,42 +1,0 @@
-.App {
-  display: flex;
-}
-
-.main {
-  padding: 40px 112px;
-  flex: 1 1 auto;
-}
-
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 52px;
-}
-
-h1 {
-  font-family: var(--font-primary);
-  font-size: var(--heading-size-md);
-  font-weight: var(--font-weight-regular);
-  color: black;
-  margin: 0 0 7px 0;
-}
-
-.name {
-  font-weight: var(--font-weight-bold);
-}
-
-.description {
-  font-family: var(--font-primary);
-  font-size: var(--body-size-lg);
-  color: var(--text-primary-color);
-  padding: 0;
-  margin: 0 0 63px 0;
-  /* line-height: 39px; */
-}
-
-.notes-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(264px, 1fr));
-  gap: 36px;
-}
