+++ src/pages/login/Login.jsx
@@ -1,0 +1,9 @@
+function Login() {
+  return (
+    <div>
+      <h1>Login</h1>
+    </div>
+  );
+}
+
+export default Login;
