+++ src/pages/signup/Signup.jsx
@@ -1,0 +1,9 @@
+function Signup() {
+  return (
+    <div>
+      <h1>Signup</h1>
+    </div>
+  );
+}
+
+export default Signup;
